Split apps into blueprints and moved main() into cicero.main.
