+++ cicero/git.py
@@ -1,20 +1,15 @@
 import flask
 import urllib
-
-import paths
 from .images import fix_images
 
-app = flask.Flask('Cicero',
-                  template_folder=paths.TEMPLATE_FOLDER,
-                  static_folder=paths.STATIC_FOLDER)
+blueprint = flask.Blueprint('git', __name__)
 
-
-@app.route('/')
+@blueprint.route('/')
 def home():
     return flask.render_template('index.html')
 
 
-@app.route('/v1/github/<user_name>/<repo_name>/<branch_name>/<file_name>/remark/')
+@blueprint.route('/v1/github/<user_name>/<repo_name>/<branch_name>/<file_name>/remark/')
 def talk(user_name, repo_name, branch_name, file_name):
     try:
         f = urllib.urlopen("https://raw.githubusercontent.com/%s/%s/%s/%s" % (user_name, repo_name, branch_name, file_name))
@@ -26,8 +21,3 @@ def talk(user_name, repo_name, branch_name, file_name):
         return flask.render_template('slides.html', markdown=''.join(fix_images(markdown, prefix)))
     except IOError:
         return flask.render_template('404.html')
-
-
-@app.errorhandler(404)
-def page_not_found(e):
-    return flask.render_template('404.html'), 404
