--- cicero/paths.py
@@ -1,6 +1,0 @@
-import os
-
-_HERE = os.path.dirname(__file__)
-
-TEMPLATE_FOLDER = os.path.join(_HERE, 'templates')
-STATIC_FOLDER = os.path.join(_HERE, 'static')
